@@ -14,40 +14,37 @@ dependencies:
     sdk: flutter
 
   # UI section
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
   provider: ^5.0.0
-  google_maps_flutter: ^2.0.1
-  flutter_polyline_points: ^0.2.4
-  flutter_webview_plugin: ^0.3.11
-  auto_size_text: ^2.1.0
-  flare_flutter: ^2.0.6
-  sliding_up_panel: ^1.0.2
+  google_maps_flutter: ^2.0.3
+  flutter_polyline_points: ^1.0.0
+  webview_flutter: ^2.0.4
+  auto_size_text: ^3.0.0-nullsafety.0
+  flare_flutter: ^3.0.0
+  sliding_up_panel: ^2.0.0+1
 
   # Logic section
-  bloc: ^7.0.0-nullsafety.3
-  hive: ^2.0.0
+  bloc: ^7.0.0
+  hive: ^2.0.4
   equatable: ^2.0.0
   rxdart: ^0.26.0
-  get_it: ^6.0.0
-  mockito: ^5.0.0
+  get_it: ^7.1.2
+  mockito: ^5.0.7
   package_info: ^2.0.0
   device_info: ^2.0.0
-  connectivity: ^3.0.2
-  dio: ^4.0.0-beta6
-  permission_handler: ^6.0.1
-  geolocator: ^7.0.1
+  connectivity: ^3.0.3
+  dio: ^4.0.0
+  permission_handler: ^7.1.0
+  geolocator: ^7.0.3
   path_provider: ^2.0.1
-  just_audio: ^0.7.0
-  audio_service: ^0.17.0-nullsafety.0
+  just_audio: ^0.7.4
+  audio_service: ^0.17.0
   url_launcher: ^6.0.2
-
-dependency_overrides:
-  http: ^0.13.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.11.5
+  build_runner: ^2.0.2
 
 flutter:
   uses-material-design: true
